+++ packages/article-summary/__tests__/article-summary-helper.js
@@ -9,9 +9,21 @@ import reviewFixture from "../fixtures/review.json";
 import blankFixture from "../fixtures/blank.json";
 
 export default ArticleSummary => {
-  it("renders an article-summary component with multiple paragraphs", () => {
-    articleMultiFixture.date = new Date("2017-07-01T14:32:00.000Z");
+  const realIntl = Intl;
 
+  beforeEach(() => {
+    global.Intl = {
+      DateTimeFormat: () => ({
+        resolvedOptions: () => ({ timeZone: "Europe/London" })
+      })
+    };
+  });
+
+  afterEach(() => {
+    global.Intl = realIntl;
+  });
+
+  it("renders an article-summary component with multiple paragraphs", () => {
     const tree = renderer
       .create(<ArticleSummary {...articleMultiFixture} />)
       .toJSON();
@@ -20,8 +32,6 @@ export default ArticleSummary => {
   });
 
   it("renders an article-summary component with a single paragraph", () => {
-    articleSingleFixture.date = new Date("2017-07-01T14:32:00.000Z");
-
     const tree = renderer
       .create(<ArticleSummary {...articleSingleFixture} />)
       .toJSON();
@@ -30,8 +40,6 @@ export default ArticleSummary => {
   });
 
   it("renders an article-summary component with content including line breaks", () => {
-    reviewFixture.date = new Date("2017-07-01T14:32:00.000Z");
-
     const tree = renderer
       .create(<ArticleSummary {...reviewFixture} />)
       .toJSON();
@@ -40,16 +48,12 @@ export default ArticleSummary => {
   });
 
   it("renders an article-summary component with headline and no content", () => {
-    blankFixture.date = new Date("2017-07-01T14:32:00.000Z");
-
     const tree = renderer.create(<ArticleSummary {...blankFixture} />).toJSON();
 
     expect(tree).toMatchSnapshot();
   });
 
   it("renders an article-summary component with empty content at the end trimmed", () => {
-    emptyParagraphFixture.date = new Date("2017-07-01T14:32:00.000Z");
-
     const tree = renderer
       .create(<ArticleSummary {...emptyParagraphFixture} />)
       .toJSON();