+++ packages/article/article-body/article-body-row.js
@@ -5,6 +5,7 @@ import { renderTrees } from "@times-components/markup";
 import ArticleImage from "@times-components/article-image";
 import PullQuote from "@times-components/pull-quote";
 import BodyParagraph from "./article-body-paragraph";
+import ArticleLink from "./article-link";
 
 const ArticleRow = ({ content: { data, index } }) =>
   renderTrees([data], {
@@ -42,6 +43,18 @@ const ArticleRow = ({ content: { data, index } }) =>
             twitter={twitter}
           />
         </View>
+      );
+    },
+    link(key, attributes, children) {
+      return (
+        <ArticleLink
+          key={index}
+          uuid={index}
+          onPress={() => {}}
+          url={attributes.href}
+        >
+          {children}
+        </ArticleLink>
       );
     }
   });