+++ packages/date-publication/__tests__/date-publication.test.js
@@ -3,45 +3,51 @@
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
-import Enzyme, { shallow } from "enzyme";
-import React16Adapter from "enzyme-adapter-react-16";
-import format from "date-fns/format";
 import DatePublication from "../date-publication";
 
-Enzyme.configure({ adapter: new React16Adapter() });
+function gmtTests(userTimezone) {
+  const realIntl = Intl;
 
-describe("Date Publication test", () => {
-  const props = {
-    date: new Date("2017-07-01T14:32:00.000Z"),
-    publication: "TIMES"
-  };
-
-  it("renders a DatePublication component with Times publication and Relevant date", () => {
-    const tree = renderer.create(<DatePublication {...props} />).toJSON();
-    expect(tree).toMatchSnapshot();
+  beforeEach(() => {
+    global.Intl = {
+      DateTimeFormat: () => ({
+        resolvedOptions: () => ({ timeZone: userTimezone })
+      })
+    };
   });
 
-  it("renders a DatePublication component with Sunday Times publication and Relevant date", () => {
-    const sundayTimesProps = {
-      ...props,
-      publication: "SUNDAYTIMES"
-    };
+  afterEach(() => {
+    global.Intl = realIntl;
+  });
+
+  it("renders a DatePublication component with Times publication and relevant GMT date", () => {
     const tree = renderer
-      .create(<DatePublication {...sundayTimesProps} />)
+      .create(
+        <DatePublication date="2017-01-01T14:32:00.000Z" publication="TIMES" />
+      )
       .toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it("date should follow the correct format", () => {
-    const component = shallow(<DatePublication {...props} />);
-    expect(component.text()).toContain(format(props.date, "dddd MMMM DD YYYY"));
+  it("renders a DatePublication component with default publication and relevant GMT date", () => {
+    const tree = renderer
+      .create(<DatePublication date="2017-01-01T14:32:00.000Z" />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
   });
 
-  it("without providing a publication, The Times will be set as default", () => {
-    const removePubProps = props;
-    delete removePubProps.publication;
-
-    const component = shallow(<DatePublication {...removePubProps} />);
-    expect(component.text()).toContain("The Times");
+  it("renders a DatePublication component with default publication and relevant BST date", () => {
+    const tree = renderer
+      .create(<DatePublication date="2017-07-01T14:32:00.000Z" />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
   });
+}
+
+describe("when the user has the same time zone as London time zone", () => {
+  gmtTests("Europe/London");
+});
+
+describe("when the user has a different time zone than London time zone", () => {
+  gmtTests("Europe/Kiev");
 });