+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.9"></a>
+## [0.7.9](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.8...@times-components/error-view@0.7.9) (2018-01-15)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.7.8"></a>
 ## [0.7.8](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.7...@times-components/error-view@0.7.8) (2018-01-08)